reorganising backend script
