+++ deployment/plantgame-deployment-backend.yml
@@ -14,5 +14,5 @@ spec:
     spec:
       containers:
       - name: plantgame-backend-container
-        image: johanneshoelker/plantgame-backend:0.0.6
+        image: johanneshoelker/plantgame-backend:0.0.7
         command: ["sleep", "infinity"]
